+++ modules/core/src/iterators/make-stream/make-dom-stream.ts
@@ -24,7 +24,6 @@ export function makeStream<ArrayBuffer>(
       // Create a byte stream (enables `Response(stream).arrayBuffer()`)
       // Only supported on Chrome
       // See: https://developer.mozilla.org/en-US/docs/Web/API/ReadableByteStreamController
-      // ts-expect-error Only Chrome supports byte streams
       type: 'bytes',
 
       async pull(controller) {