+++ modules/wms/src/bundle.ts
@@ -1,0 +1,4 @@
+// @ts-nocheck
+const moduleExports = require('./index');
+globalThis.loaders = globalThis.loaders || {};
+module.exports = Object.assign(globalThis.loaders, moduleExports);