+++ modules/wms/src/lib/parse-wms-capabilities.ts
@@ -1,0 +1,7 @@
+// loaders.gl, MIT license
+
+import {XMLLoader} from '@loaders.gl/xml';
+
+export function parseWMSCapabilities(text, options) {
+  return XMLLoader.parseTextSync(text, options);
+}